+++ src/main/groovy/com/dispatchlabs/io/testing/common/contracts/DefaultSampleContract.groovy
@@ -1,0 +1,49 @@
+package com.dispatchlabs.io.testing.common.contracts
+
+class DefaultSampleContract {
+    //Default Sample Contract
+    /* Contract
+    pragma solidity ^0.4.0;
+contract TestContract {
+
+    struct ComplexStruct {
+        uint var1;
+        bool var2;
+        uint8 var3;
+        string var4;
+    }
+
+    string public var5;
+    ComplexStruct public var6;
+
+    constructor() public {
+        var5 = "aaaaaaaaaaaaa";
+        var6.var1 = 0;
+        var6.var2 = false;
+        var6.var3 = 1;
+        var6.var4 = "bbbbbbbbbbb";
+    }
+
+    function setVar5(string value) public {
+        var5 = value;
+    }
+
+    function setVar6Var4(string value) public {
+        var6.var4 = value;
+    }
+
+    function incVar6Var1() public {
+        var6.var1++;
+    }
+
+    function getVar5() public view returns (string) {
+        return var5;
+    }
+}
+     */
+
+    public static String defaultSample =
+        '''608060405234801561001057600080fd5b506040805190810160405280600d81526020017f61616161616161616161616161000000000000000000000000000000000000008152506000908051906020019061005c9291906100f7565b50600060016000018190555060006001800160006101000a81548160ff02191690831515021790555060018060010160016101000a81548160ff021916908360ff1602179055506040805190810160405280600b81526020017f6262626262626262626262000000000000000000000000000000000000000000815250600160020190805190602001906100f19291906100f7565b5061019c565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061013857805160ff1916838001178555610166565b82800160010185558215610166579182015b8281111561016557825182559160200191906001019061014a565b5b5090506101739190610177565b5090565b61019991905b8082111561019557600081600090555060010161017d565b5090565b90565b610664806101ab6000396000f300608060405260043610610078576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff16806333e538e91461007d57806334e45f531461010d5780633a458b1f1461017657806378d8866e1461022557806379af6473146102b5578063cb69e300146102cc575b600080fd5b34801561008957600080fd5b50610092610335565b6040518080602001828103825283818151815260200191508051906020019080838360005b838110156100d25780820151818401526020810190506100b7565b50505050905090810190601f1680156100ff5780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b34801561011957600080fd5b50610174600480360381019080803590602001908201803590602001908080601f01602080910402602001604051908101604052809392919081815260200183838082843782019150505050505091929192905050506103d7565b005b34801561018257600080fd5b5061018b6103f4565b60405180858152602001841515151581526020018360ff1660ff16815260200180602001828103825283818151815260200191508051906020019080838360005b838110156101e75780820151818401526020810190506101cc565b50505050905090810190601f1680156102145780820380516001836020036101000a031916815260200191505b509550505050505060405180910390f35b34801561023157600080fd5b5061023a6104c4565b6040518080602001828103825283818151815260200191508051906020019080838360005b8381101561027a57808201518184015260208101905061025f565b50505050905090810190601f1680156102a75780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b3480156102c157600080fd5b506102ca610562565b005b3480156102d857600080fd5b50610333600480360381019080803590602001908201803590602001908080601f0160208091040260200160405190810160405280939291908181526020018383808284378201915050505050509192919290505050610579565b005b606060008054600181600116156101000203166002900480601f0160208091040260200160405190810160405280929190818152602001828054600181600116156101000203166002900480156103cd5780601f106103a2576101008083540402835291602001916103cd565b820191906000526020600020905b8154815290600101906020018083116103b057829003601f168201915b5050505050905090565b80600160020190805190602001906103f0929190610593565b5050565b60018060000154908060010160009054906101000a900460ff16908060010160019054906101000a900460ff1690806002018054600181600116156101000203166002900480601f0160208091040260200160405190810160405280929190818152602001828054600181600116156101000203166002900480156104ba5780601f1061048f576101008083540402835291602001916104ba565b820191906000526020600020905b81548152906001019060200180831161049d57829003601f168201915b5050505050905084565b60008054600181600116156101000203166002900480601f01602080910402602001604051908101604052809291908181526020018280546001816001161561010002031660029004801561055a5780601f1061052f5761010080835404028352916020019161055a565b820191906000526020600020905b81548152906001019060200180831161053d57829003601f168201915b505050505081565b600160000160008154809291906001019190505550565b806000908051906020019061058f929190610593565b5050565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f106105d457805160ff1916838001178555610602565b82800160010185558215610602579182015b828111156106015782518255916020019190600101906105e6565b5b50905061060f9190610613565b5090565b61063591905b80821115610631576000816000905550600101610619565b5090565b905600a165627a7a7230582026a289af0b033267e3fc13869c446b0552e2c62b9b3fa46bc626be2c683528680029'''
+    public static String defaultSampleABI =
+        '''5b7b0a2020202022636f6e7374616e74223a20747275652c0a2020202022696e70757473223a205b5d2c0a20202020226e616d65223a202267657456617235222c0a20202020226f757470757473223a205b0a2020202020207b0a2020202020202020226e616d65223a2022222c0a20202020202020202274797065223a2022737472696e67220a2020202020207d0a202020205d2c0a202020202270617961626c65223a2066616c73652c0a202020202273746174654d75746162696c697479223a202276696577222c0a202020202274797065223a202266756e6374696f6e220a20207d2c0a20207b0a2020202022636f6e7374616e74223a2066616c73652c0a2020202022696e70757473223a205b0a2020202020207b0a2020202020202020226e616d65223a202276616c7565222c0a20202020202020202274797065223a2022737472696e67220a2020202020207d0a202020205d2c0a20202020226e616d65223a20227365745661723656617234222c0a20202020226f757470757473223a205b5d2c0a202020202270617961626c65223a2066616c73652c0a202020202273746174654d75746162696c697479223a20226e6f6e70617961626c65222c0a202020202274797065223a202266756e6374696f6e220a20207d2c0a20207b0a2020202022636f6e7374616e74223a20747275652c0a2020202022696e70757473223a205b5d2c0a20202020226e616d65223a202276617236222c0a20202020226f757470757473223a205b0a2020202020207b0a2020202020202020226e616d65223a202276617231222c0a20202020202020202274797065223a202275696e74323536220a2020202020207d2c0a2020202020207b0a2020202020202020226e616d65223a202276617232222c0a20202020202020202274797065223a2022626f6f6c220a2020202020207d2c0a2020202020207b0a2020202020202020226e616d65223a202276617233222c0a20202020202020202274797065223a202275696e7438220a2020202020207d2c0a2020202020207b0a2020202020202020226e616d65223a202276617234222c0a20202020202020202274797065223a2022737472696e67220a2020202020207d0a202020205d2c0a202020202270617961626c65223a2066616c73652c0a202020202273746174654d75746162696c697479223a202276696577222c0a202020202274797065223a202266756e6374696f6e220a20207d2c0a20207b0a2020202022636f6e7374616e74223a20747275652c0a2020202022696e70757473223a205b5d2c0a20202020226e616d65223a202276617235222c0a20202020226f757470757473223a205b0a2020202020207b0a2020202020202020226e616d65223a2022222c0a20202020202020202274797065223a2022737472696e67220a2020202020207d0a202020205d2c0a202020202270617961626c65223a2066616c73652c0a202020202273746174654d75746162696c697479223a202276696577222c0a202020202274797065223a202266756e6374696f6e220a20207d2c0a20207b0a2020202022636f6e7374616e74223a2066616c73652c0a2020202022696e70757473223a205b5d2c0a20202020226e616d65223a2022696e635661723656617231222c0a20202020226f757470757473223a205b5d2c0a202020202270617961626c65223a2066616c73652c0a202020202273746174654d75746162696c697479223a20226e6f6e70617961626c65222c0a202020202274797065223a202266756e6374696f6e220a20207d2c0a20207b0a2020202022636f6e7374616e74223a2066616c73652c0a2020202022696e70757473223a205b0a2020202020207b0a2020202020202020226e616d65223a202276616c7565222c0a20202020202020202274797065223a2022737472696e67220a2020202020207d0a202020205d2c0a20202020226e616d65223a202273657456617235222c0a20202020226f757470757473223a205b5d2c0a202020202270617961626c65223a2066616c73652c0a202020202273746174654d75746162696c697479223a20226e6f6e70617961626c65222c0a202020202274797065223a202266756e6374696f6e220a20207d2c0a20207b0a2020202022696e70757473223a205b5d2c0a202020202270617961626c65223a2066616c73652c0a202020202273746174654d75746162696c697479223a20226e6f6e70617961626c65222c0a202020202274797065223a2022636f6e7374727563746f72220a20207d0a5d'''
+}
