+++ src/test/groovy/api/transactions/RegressionTransactions.groovy
@@ -1,12 +1,10 @@
 package api.transactions
 
-import com.dispatchlabs.io.testing.common.Contracts
 import com.dispatchlabs.io.testing.common.NodeSetup
 import org.testng.annotations.BeforeMethod
 import org.testng.annotations.Test
 
 import static com.dispatchlabs.io.testing.common.APIs.*
-import static org.hamcrest.Matchers.equalTo
 
 class RegressionTransactions {
     def allNodes
@@ -102,160 +100,7 @@ class RegressionTransactions {
         verifyConsensusForAccount Nodes:allNodes.Delegates, ID:allNodes.Delegates.Delegate0.address, Status: "Ok", Balance: 999
     }
 
-    @Test(description="Smart Contract Smoke",groups = ["smoke", "smart contract"])
-    public void transactionRegression6_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "setVar5",Params: ["5555"]
-        waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "getVar5",Params: []
-        def getID = response.then().extract().path("id")
-        waitForTransactionStatus ID:getID ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        verifyStatusForTransaction(Nodes:[allNodes.Delegates.Delegate0],ID:getID,ContractResult:"5555")
-    }
 
-    @Test(description="Pass invalid value for parameter type: string",groups = ["smart contract"])
-    public void transactionRegression7_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "setVar5",Params: "5555"
-        response.then().assertThat().body("type",equalTo("InvalidTransaction"))
-    }
-
-    @Test(description="Pass invalid value for parameter type: hash",groups = ["smart contract"])
-    public void transactionRegression8_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "setVar5",Params: ["value":"5555"]
-        response.then().assertThat().body("type",equalTo("InvalidTransaction"))
-    }
-
-    @Test(description="Pass invalid value for parameter type: integer",groups = ["smart contract"])
-    public void transactionRegression9_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "setVar5",Params: 5555
-        response.then().assertThat().body("type",equalTo("InvalidTransaction"))
-    }
-
-    @Test(description="Pass invalid value for parameter type: boolean",groups = ["smart contract"])
-    public void transactionRegression10_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "setVar5",Params: false
-        response.then().assertThat().body("type",equalTo("InvalidTransaction"))
-    }
-
-    @Test(description="Contract returns no value",groups = ["smart contract"])
-    public void transactionRegression11_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "setVar5",Params: ["10"]
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        response.then().assertThat().body("contractAddress",equalTo(contractAddress))
-    }
-
-    @Test(description="Negative: Pass only some parameter values for methods",groups = ["smart contract"])
-    public void transactionRegression12_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "setVar5",Params: []
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        response.then().assertThat().body("type",equalTo("InvalidTransaction"))
-    }
-
-    @Test(description="Deploy exact same contract",groups = ["smart contract"])
-    public void transactionRegression13_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress1 = response.then().extract().path("contractAddress")
-
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress2 = response.then().extract().path("contractAddress")
-
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress1, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "setVar5",Params: ["1234"]
-        waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress2, PrivateKey:"Genesis",Type:2,
-                ABI: Contracts.defaultSampleABI,
-                Method: "getVar5"
-        def getID = response.then().extract().path("id")
-        waitForTransactionStatus ID:getID ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        verifyStatusForTransaction(Nodes:[allNodes.Delegates.Delegate0],ID:getID,ContractResult:"aaaaaaaaaaaaa")
-    }
-
-    @Test(description="Negative: Empty ABI on contract method call",groups = ["smart contract"])
-    public void transactionRegression14_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                Method: "setVar5",Params: ["10"]
-        response.then().assertThat().body("type",equalTo("InvalidTransaction"))
-    }
-
-    @Test(description="Negative: Invalid values for ABI",groups = ["smart contract"])
-    public void transactionRegression15_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample
-        response = waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        def contractAddress = response.then().extract().path("contractAddress")
-        response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:contractAddress, PrivateKey:"Genesis",Type:2,
-                ABI: "adfsdfsdfsfsdffffffffffffffffffffffffff",
-                Method: "setVar5",Params: ["10"]
-        waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-        response.then().assertThat().body("type",equalTo("InvalidTransaction"))
-    }
-
-    @Test(description="Negative: Empty smart contract",groups = ["smart contract"])
-    public void transactionRegression16_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:""
-        waitForTransactionStatus ID:response.then().extract().path("id") ,Node:allNodes.Delegates.Delegate0, Status: "Ok", Timeout: 10
-    }
-
-    @Test(description="Negative: Partial bytecode of valid contract",groups = ["smart contract"])
-    public void transactionRegression17_SmartContract(){
-        def response = sendTransaction Node:allNodes.Delegates.Delegate0, Value:0,From:"Genesis",To:"", PrivateKey:"Genesis",Type:1,
-                Code:Contracts.defaultSample.substring(0,Contracts.defaultSample.size()-10)
-        response.then().assertThat().body("type",equalTo("InvalidTransaction"))
-    }
 
     /*
     @Test(description="Decimal point tokens",groups = ["smoke", "transactions"])
